Create README.md
MONTY:
0x19. C - Stacks, Queues - LIFO, FIFO PROJECT WRITTEN BY IYENAGBE RAYMOND OGEYANOR AND GABRIEL IFEANTI

Tests
We strongly encourage you to work all together on a set of tests

The Monty language
Monty 0.98 is a scripting language that is first compiled into Monty byte codes (Just like Python). It relies on a unique stack, with specific instructions to manipulate it. The goal of this project is to create an interpreter for Monty ByteCodes files.
Monty byte code files

Files containing Monty byte codes usually have the .m extension. Most of the industry uses this standard but it is not required by the specification of the language. There is not more than one instruction per line. There can be any number of spaces before or after the opcode and its argument: